feat(MAM-24): 필터링 바텀시트 컴포넌트 리팩토링 및 새로운 기능 추가
- 기존의 가격대 및 카테고리 필터링 바텀시트를 통합하여 use필터링_바텀시트 훅으로 관리
- 필터링 바텀시트에 적용 버튼 및 리셋 기능 추가
- 카테고리 필터링 기능을 위한 새로운 컴포넌트 및 스타일 추가
- 불필요한 코드 제거 및 가독성 향상
